archive_originals/requirements.txt
anthropic>=0.40.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
